Rename aya_kun to aya_gen8
