+++ src/handycon/devices.py
@@ -19,7 +19,7 @@ import handycon.handhelds.aya_gen4 as aya_gen4
 import handycon.handhelds.aya_gen5 as aya_gen5
 import handycon.handhelds.aya_gen6 as aya_gen6
 import handycon.handhelds.aya_gen7 as aya_gen7
-import handycon.handhelds.aya_kun as aya_kun
+import handycon.handhelds.aya_gen8 as aya_gen8
 import handycon.handhelds.ayn_gen1 as ayn_gen1
 import handycon.handhelds.ayn_gen2 as ayn_gen2
 import handycon.handhelds.ayn_gen3 as ayn_gen3
@@ -268,8 +268,8 @@ async def capture_keyboard_events():
                             await aya_gen6.process_event(seed_event, active_keys)
                         case "AYA_GEN7":
                             await aya_gen7.process_event(seed_event, active_keys)
-                        case "AYA_KUN":
-                            await aya_kun.process_event(seed_event, active_keys)
+                        case "AYA_GEN8":
+                            await aya_gen8.process_event(seed_event, active_keys)
                         case "AYN_GEN1":
                             await ayn_gen1.process_event(seed_event, active_keys)
                         case "AYN_GEN2":
